refactor: extract envelopes view builder and simplify data provider
Extract buildEnvelopesViewData to dedicated file, reducing data provider
to a thin facade. Use shared constants from budget-table-constants.
